docufraud/requirements.txt
# Integrated Tampering Detection Requirements

# Core dependencies
numpy>=1.24.0
Pillow>=10.0.0
opencv-python>=4.8.0

# Image processing and forensics
scikit-image>=0.21.0
imagehash>=4.3.1

# PDF processing
PyMuPDF>=1.23.0

# OCR (optional but recommended)
pytesseract>=0.3.10

# Azure OpenAI
openai>=1.12.0

# Utilities
python-dateutil>=2.8.2

# Domain and network verification
dnspython>=2.4.0
requests>=2.31.0

# Document processing
python-docx>=1.1.0
